+++ Gems/CsvSpawner/Code/Source/CsvSpawner/CsvSpawnerComponent.h
@@ -10,7 +10,7 @@
 
 #pragma once
 
-#include "CsvSpawnerUtils.h"
+#include <CsvSpawner/CsvSpawnerUtils.h>
 
 #include <AzCore/Component/Component.h>
 #include <AzCore/Component/TickBus.h>
