+++ Gems/CsvSpawner/Code/Source/CsvSpawner/CsvSpawnerUtils.cpp
@@ -9,7 +9,7 @@
  * file.
  */
 
-#include "CsvSpawnerUtils.h"
+#include <CsvSpawner/CsvSpawnerUtils.h>
 
 #include <AzFramework/Physics/CollisionBus.h>
 #include <CsvSpawner/CsvSpawnerInterface.h>
