+++ Gems/CsvSpawner/Code/csvspawner_api_files.cmake
@@ -2,4 +2,5 @@
 set(FILES
     Include/CsvSpawner/CsvSpawnerInterface.h
     Include/CsvSpawner/CsvSpawnerTypeIds.h
+    Include/CsvSpawner/CsvSpawnerUtils.h
 )
