+++ Gems/CsvSpawner/Code/csvspawner_private_files.cmake
@@ -3,7 +3,6 @@ set(FILES
     Source/CsvSpawnerModuleInterface.cpp
     Source/CsvSpawnerModuleInterface.h
     Source/CsvSpawner/CsvSpawnerUtils.cpp
-    Source/CsvSpawner/CsvSpawnerUtils.h
     Source/CsvSpawner/CsvSpawnerComponent.cpp
     Source/CsvSpawner/CsvSpawnerComponent.h
 )
